requirements.txt
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
torch==2.5.1
fastapi==0.115.6
uvicorn==0.32.1
jinja2==3.1.5
matplotlib==3.9.2
seaborn==0.13.2
ipykernel==6.29.5
mlflow==2.17.2
dvc[s3]==3.58.0
pathspec==0.12.1
pyyaml==6.0.2
joblib==1.4.2
pytest==8.3.3
ruff==0.7.4
